requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
langchain==0.1.7
python-dotenv==1.0.1
langchain-google-genai==1.0.1
langchainhub==0.1.15